requirements.txt
jinja2
markdown
google-generativeai
genkit-plugin-google-genai